fix(utils): do not import vtk unnecessarily
