+++ trame_rca/protocol.py
@@ -1,5 +1,47 @@
+from typing import Protocol, runtime_checkable
+from numpy.typing import NDArray
 from wslink import register as exportRpc
 from wslink.websocket import LinkProtocol
+
+
+@runtime_checkable
+class AbstractWindow(Protocol):
+    """
+    Protocol defining the interface for interacting with a remote window through RCA.
+
+    Any class matching this interface can be used as a remote window, regardless of inheritance.
+    Implementing classes must define the required methods and properties to enable window interaction.
+    """
+
+    @property
+    def img_cols_rows(self) -> tuple[NDArray, int, int]:
+        """
+        Returns a tuple containing:
+        - the window content as a NumPy array,
+        - the number of columns,
+        - and the number of rows.
+
+        Called by the scheduler to render the current window view.
+        """
+        pass
+
+    def process_resize_event(self, width: int, height: int) -> None:
+        """
+        Handle a resize event for the RCA (RenderWindowInteractor).
+
+        This method is triggered by the adapter whenever the window is resized.
+        """
+        pass
+
+    def process_interaction_event(self, event: dict) -> None:
+        """
+        Handle an interaction event from the RCA (RenderWindowInteractor).
+
+        This method is invoked by the adapter whenever an interaction event occurs.
+        Refer to the event types defined in:
+        https://github.com/Kitware/vtk-js/blob/master/Sources/Rendering/Core/RenderWindowInteractor/index.js
+        """
+        pass
 
 
 class AreaAdapter:
